requirements.txt
ply>=3.11
pywebview>=4.0
